requirements.txt
amqp==5.1.1
asgiref==3.4.1
asn1crypto==0.24.0
billiard==3.6.4.0
boto3==1.16.63
botocore==1.19.63
cached-property==1.5.2
celery==5.0.5
certifi==2021.10.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==2.1.4
Django==3.1.3
django-debug-toolbar==3.2.4
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-ratelimit==3.0.1
django-storages==1.10.1
djangorestframework==3.12.2
happybase==1.2.0
idna==2.6
importlib-metadata==4.8.3
jmespath==0.10.0
jsonfield==3.1.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
mysqlclient==2.0.3
ply==3.11
prompt-toolkit==3.0.29
pycrypto==2.6.1
PyGObject==3.26.1
python-dateutil==2.8.2
python-memcached==1.59
pytz==2022.1
pyxdg==0.25
redis==3.5.3
requests==2.18.4
s3transfer==0.3.7
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.4.2
ssh-import-id==5.7
swapper==1.3.0
thriftpy2==0.4.14
typing_extensions==4.1.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.6.0
